project/requirements.txt
python-docx==1.1.2
spacy>=3.7.0
jinja2>=3.1.4

